Fixing the Likert test to make sure it doesn't import another test and force that to run as well
