+++ src/layout/Group/GroupContainer.test.tsx
@@ -7,11 +7,10 @@ import type { PayloadAction } from '@reduxjs/toolkit';
 import { getFormLayoutGroupMock } from 'src/__mocks__/formLayoutGroupMock';
 import { getInitialStateMock } from 'src/__mocks__/initialStateMock';
 import { FormLayoutActions } from 'src/features/layout/formLayoutSlice';
-import { GroupContainer } from 'src/layout/Group/GroupContainer';
+import { GroupContainerTester } from 'src/layout/Group/GroupContainerTestUtills';
 import { setupStore } from 'src/redux/store';
 import { mockMediaQuery, renderWithProviders } from 'src/testUtils';
 import { Triggers } from 'src/types';
-import { useResolvedNode } from 'src/utils/layout/ExprContext';
 import type { ExprUnresolved } from 'src/features/expressions/types';
 import type {
   IUpdateRepeatingGroupsEditIndex,
@@ -24,15 +23,6 @@ const mockContainer = getFormLayoutGroupMock();
 
 interface IRender {
   container?: ExprUnresolved<ILayoutGroup>;
-}
-
-export function GroupContainerTester(props: { id: string }) {
-  const node = useResolvedNode(props.id);
-  if (!node || !node.isRepGroup()) {
-    throw new Error(`Could not resolve node with id ${props.id}, or unexpected node type`);
-  }
-
-  return <GroupContainer node={node} />;
 }
 
 function render({ container = mockContainer }: IRender = {}) {
